requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
durationpy==0.9
exceptiongroup==1.2.2
fastapi==0.115.12
Flask==3.1.0
Flask-SocketIO==5.5.1
google-auth==2.38.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kubernetes==32.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.0
pydantic_core==2.33.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tabulate==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
